fix paths for build package
